Break up the AwsExplorerNodes (#1091)
* No functional changes were made
