+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/cloudformation/CloudFormationExplorerNodes.kt
@@ -13,13 +13,13 @@ import software.amazon.awssdk.services.lambda.LambdaClient
 import software.aws.toolkits.jetbrains.core.AwsClientManager
 import software.aws.toolkits.jetbrains.core.DeleteResourceAction
 import software.aws.toolkits.jetbrains.core.awsClient
-import software.aws.toolkits.jetbrains.core.explorer.AwsExplorerEmptyNode
-import software.aws.toolkits.jetbrains.core.explorer.AwsExplorerErrorNode
-import software.aws.toolkits.jetbrains.core.explorer.AwsExplorerNode
-import software.aws.toolkits.jetbrains.core.explorer.AwsExplorerResourceNode
-import software.aws.toolkits.jetbrains.core.explorer.AwsExplorerServiceRootNode
-import software.aws.toolkits.jetbrains.core.explorer.AwsNodeAlwaysExpandable
-import software.aws.toolkits.jetbrains.core.explorer.AwsTruncatedResultNode
+import software.aws.toolkits.jetbrains.core.explorer.nodes.AwsExplorerEmptyNode
+import software.aws.toolkits.jetbrains.core.explorer.nodes.AwsExplorerErrorNode
+import software.aws.toolkits.jetbrains.core.explorer.nodes.AwsExplorerNode
+import software.aws.toolkits.jetbrains.core.explorer.nodes.AwsExplorerResourceNode
+import software.aws.toolkits.jetbrains.core.explorer.nodes.AwsExplorerServiceRootNode
+import software.aws.toolkits.jetbrains.core.explorer.nodes.AwsNodeAlwaysExpandable
+import software.aws.toolkits.jetbrains.core.explorer.nodes.AwsTruncatedResultNode
 import software.aws.toolkits.jetbrains.services.lambda.LambdaFunctionNode
 import software.aws.toolkits.jetbrains.services.lambda.toDataClass
 import software.aws.toolkits.jetbrains.utils.TaggingResourceType
@@ -46,7 +46,12 @@ class CloudFormationServiceNode(project: Project) : AwsExplorerServiceRootNode(p
     }
 
     private fun paginationNodeIfRequired(nextToken: String?) = when {
-        nextToken != null -> listOf(AwsTruncatedResultNode(this, nextToken))
+        nextToken != null -> listOf(
+            AwsTruncatedResultNode(
+                this,
+                nextToken
+            )
+        )
         else -> emptyList()
     }
 
@@ -70,7 +75,12 @@ class CloudFormationStackNode(project: Project, val stackName: String, private v
      * CloudFormation Stack Nodes do not immediately query for stack resources.
      * We wait until node will be expanded before querying, reducing risk of triggering TPS limits.
      */
-    private val noResourcesChildren: Collection<AbstractTreeNode<Any>> = listOf(AwsExplorerEmptyNode(project, message("explorer.stack.no.serverless.resources"))).filterIsInstance<AbstractTreeNode<Any>>()
+    private val noResourcesChildren: Collection<AbstractTreeNode<Any>> = listOf(
+        AwsExplorerEmptyNode(
+            project,
+            message("explorer.stack.no.serverless.resources")
+        )
+    ).filterIsInstance<AbstractTreeNode<Any>>()
     private var cachedChildren: Collection<AbstractTreeNode<Any>> = emptyList()
 
     var isChildCacheInInitialState: Boolean = true
